add middleware for permission
