+++ errors.go
@@ -1,0 +1,7 @@
+package auth
+
+import (
+	"errors"
+)
+
+var ErrInvalidUser = errors.New("Invalid User")
